Refactor Serializers and API Views into separate files
